# backend/src/main/resources/static/api-docs.yaml
openapi: 3.0.0
info:
  title: Springyield API
  version: v0.1
servers:
  - url: http://springyield-springboot:8080
  - url: https://api.springyield.stefvisser.com
paths:
  /api/user/{userId}/update:
    put:
      tags:
        - user-controller
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/user/{userId}/approve:
    put:
      tags:
        - user-controller
      operationId: approveUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserApprovalDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/user/{userId}:
    get:
      tags:
        - user-controller
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/user/search:
    get:
      tags:
        - user-controller
      operationId: getUserByName
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum:
              - UNAPPROVED
              - APPROVED
              - EMPLOYEE
              - ADMIN
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/user/{userId}/delete:
    delete:
      tags:
        - user-controller
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
    UserUpdateDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        bsnNumber:
          type: integer
          format: int32
        phoneNumber:
          type: string
        role:
          type: string
          enum:
            - UNAPPROVED
            - APPROVED
            - EMPLOYEE
            - ADMIN
        password:
          type: string
    UserApprovalDTO:
      type: object
      properties:
        dailyLimit:
          type: number
        absoluteLimit:
          type: number
...